external-scraper-api/src/scraper/certification_scraper.py
import requests
from bs4 import BeautifulSoup
from typing import List, Dict, Optional
import logging
from datetime import datetime

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class CertificationScraper:
    def __init__(self, base_url: str):
        self.base_url = base_url
        self.session = requests.Session()

    async def get_certifications(self, lss_id: str) -> List[Dict]:
        """
        Scrape certifications for a given LSS ID.
        This is a mock implementation. In production, this would scrape the actual certification data.
        """
        try:
            # Mock implementation
            # In production, this would make actual HTTP requests and parse the response
            logger.info(f"Fetching certifications for LSS ID: {lss_id}")
            
            # Simulate network delay
            await asyncio.sleep(1)
            
            # Return mock data
            return [
                {
                    "id": f"cert-{lss_id}-001",
                    "name": "Advanced Leadership",
                    "issuer": "Leadership Institute",
                    "date": "2022-01-15",
                    "expiry_date": "2025-01-15"
                },
                {
                    "id": f"cert-{lss_id}-002",
                    "name": "Professional Coaching",
                    "issuer": "International Coaching Federation",
                    "date": "2021-06-20",
                    "expiry_date": None
                }
            ]
        except Exception as e:
            logger.error(f"Error fetching certifications: {str(e)}")
            raise

    async def verify_certification(self, lss_id: str, certification_id: str) -> Dict:
        """
        Verify a specific certification.
        This is a mock implementation. In production, this would verify against the actual certification database.
        """
        try:
            # Mock implementation
            # In production, this would make actual verification requests
            logger.info(f"Verifying certification {certification_id} for LSS ID: {lss_id}")
            
            # Simulate network delay
            await asyncio.sleep(0.5)
            
            # Return mock verification result
            return {
                "isValid": True,
                "details": {
                    "verified_at": datetime.utcnow().isoformat(),
                    "verification_method": "database_check",
                    "certification_id": certification_id,
                    "lss_id": lss_id
                }
            }
        except Exception as e:
            logger.error(f"Error verifying certification: {str(e)}")
            raise

    def _parse_certification_data(self, html_content: str) -> List[Dict]:
        """
        Parse HTML content to extract certification data.
        This is a placeholder for the actual parsing logic.
        """
        # This would contain the actual parsing logic using BeautifulSoup
        # For now, return empty list as this is a mock implementation
        return []

    def _validate_certification(self, certification: Dict) -> bool:
        """
        Validate certification data.
        This is a placeholder for the actual validation logic.
        """
        # This would contain the actual validation logic
        # For now, return True as this is a mock implementation
        return True 